# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEO Asset Management API",
        "description": "REST API for VEO Asset Management System mobile application",
        "contact": {
            "email": "support@veoset.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://veoset.test",
            "description": "VEO Asset Management API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard overview",
                "operationId": "aa916ea9a152b88ff8d02de260dd3390",
                "responses": {
                    "200": {
                        "description": "Dashboard overview data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/statistics": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard statistics",
                "operationId": "addf8c82a321a06898e953647b288fd4",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period",
                        "schema": {
                            "type": "string",
                            "default": "week",
                            "enum": [
                                "today",
                                "week",
                                "month"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard statistics"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}